Add RootedRc::fast_clone
I noticed that RootedRefCell had a bigger performance advantage than
RootedRc and suspect that's because it validates using a reference to
the lock instead of accessing the thread-local tracker.

`fast_clone` does the same. It's a bit less convenient than `clone`
since it requires a reference to the lock, and doesn't work for
recursive clones, but it substantially faster.