Simplify implementation
Removed thread-locals to track whether the root lock was held turned out
to significantly hurt performance, as well as complicating the
implementation. APIs that require the root lock to be held now all
require explicitly passing a reference to the held lock instead.

Removed capability of the root lock to hold additional data, since this
also significantly complicated the implementation. A similar effect can
be achieved by bundling a `Root` and a `RootedRefCell` together.

Updated benchmarks, examples, README, etc.