RootGuard: ensure !Sync
There was a soundness bug - the guard could be accessed simultaneously
by multiple threads, invalidating the safety proofs in GuardedRc and
GuardedRefCell, which take a &RootGuard to prove that the current thread
has exclusive access to the corresponding Root.

I verified the problem by adding a test that failed miri. After fixing
the bug, I fixed the test so that it still compiles.

It'd be nice to have a regression test enforcing that the previous
version *doesn't* compile, but I don't know of a straightforward way to
do that.