Remove RootGuard
We make `Root` itself `!Sync`, which allows us to get rid of
`RootGuard`. Providing a reference to `Root` is sufficient to prove that
no other threads are simultaneously accessing any of the objects
associated with it.

It may still be useful to wrap a `Root` inside a `Mutex`, but now the
user has the flexibility to choose the type of `Mutex`, or avoid it by
sending the `Root` between threads.